extract top tests
