+++ bpaf_derive/src/lib.rs
@@ -9,6 +9,8 @@ mod kw;
 mod top;
 mod utils;
 use top::Top;
+#[cfg(test)]
+mod top_tests;
 
 /// Derive macro for bpaf command line parser
 ///
